# bench: add logging benchmarks for LogPrintLevel and LogPrintfCategory