--- a/src/bench/logging.cpp
+++ b/src/bench/logging.cpp
@@ -25,6 +25,14 @@ static void LoggingNoThreadNames(benchmark::Bench& bench)
 {
     Logging(bench, {"-logthreadnames=0"}, [] { LogPrintf("%s\n", "test"); });
 }
+static void LogPrintfCategoryWithThreadNames(benchmark::Bench& bench)
+{
+    Logging(bench, {"-logthreadnames=1"}, [] { LogPrintfCategory(BCLog::NET, "%s\n", "test"); });
+}
+static void LogPrintfCategoryWithoutThreadNames(benchmark::Bench& bench)
+{
+    Logging(bench, {"-logthreadnames=0"}, [] { LogPrintfCategory(BCLog::NET, "%s\n", "test"); });
+}
 static void LoggingYoCategory(benchmark::Bench& bench)
 {
     Logging(bench, {"-logthreadnames=0", "-debug=net"}, [] { LogPrint(BCLog::NET, "%s\n", "test"); });
@@ -37,12 +45,15 @@ static void LoggingNoFile(benchmark::Bench& bench)
 {
     Logging(bench, {"-nodebuglogfile", "-debug=1"}, [] {
         LogPrintf("%s\n", "test");
+        LogPrintfCategory(BCLog::NET, "%s\n", "test");
         LogPrint(BCLog::NET, "%s\n", "test");
     });
 }
 
 BENCHMARK(LoggingYoThreadNames);
 BENCHMARK(LoggingNoThreadNames);
+BENCHMARK(LogPrintfCategoryWithThreadNames);
+BENCHMARK(LogPrintfCategoryWithoutThreadNames);
 BENCHMARK(LoggingYoCategory);
 BENCHMARK(LoggingNoCategory);
 BENCHMARK(LoggingNoFile);