--- a/src/bench/logging.cpp
+++ b/src/bench/logging.cpp
@@ -17,11 +17,11 @@ static void Logging(benchmark::Bench& bench, const std::vector<const char*>& ext
     bench.run([&] { log(); });
 }
 
-static void LoggingYoThreadNames(benchmark::Bench& bench)
+static void LogPrintfWithThreadNames(benchmark::Bench& bench)
 {
     Logging(bench, {"-logthreadnames=1"}, [] { LogPrintf("%s\n", "test"); });
 }
-static void LoggingNoThreadNames(benchmark::Bench& bench)
+static void LogPrintfWithoutThreadNames(benchmark::Bench& bench)
 {
     Logging(bench, {"-logthreadnames=0"}, [] { LogPrintf("%s\n", "test"); });
 }
@@ -43,15 +43,15 @@ static void LogPrintLevelWithoutThreadNames(benchmark::Bench& bench)
     Logging(bench, {"-logthreadnames=0"}, [] {
         LogPrintLevel(BCLog::NET, BCLog::Level::Error, "%s\n", "test"); });
 }
-static void LoggingYoCategory(benchmark::Bench& bench)
+static void LogPrintWithCategory(benchmark::Bench& bench)
 {
     Logging(bench, {"-logthreadnames=0", "-debug=net"}, [] { LogPrint(BCLog::NET, "%s\n", "test"); });
 }
-static void LoggingNoCategory(benchmark::Bench& bench)
+static void LogPrintWithoutCategory(benchmark::Bench& bench)
 {
     Logging(bench, {"-logthreadnames=0", "-debug=0"}, [] { LogPrint(BCLog::NET, "%s\n", "test"); });
 }
-static void LoggingNoFile(benchmark::Bench& bench)
+static void LogNoDebugLogFile(benchmark::Bench& bench)
 {
     Logging(bench, {"-nodebuglogfile", "-debug=1"}, [] {
         LogPrintf("%s\n", "test");
@@ -61,12 +61,12 @@ static void LoggingNoFile(benchmark::Bench& bench)
     });
 }
 
-BENCHMARK(LoggingYoThreadNames);
-BENCHMARK(LoggingNoThreadNames);
+BENCHMARK(LogPrintfWithThreadNames);
+BENCHMARK(LogPrintfWithoutThreadNames);
 BENCHMARK(LogPrintfCategoryWithThreadNames);
 BENCHMARK(LogPrintfCategoryWithoutThreadNames);
 BENCHMARK(LogPrintLevelWithThreadNames);
 BENCHMARK(LogPrintLevelWithoutThreadNames);
-BENCHMARK(LoggingYoCategory);
-BENCHMARK(LoggingNoCategory);
-BENCHMARK(LoggingNoFile);
+BENCHMARK(LogPrintWithCategory);
+BENCHMARK(LogPrintWithoutCategory);
+BENCHMARK(LogNoDebugLogFile);