@jonatack 2022-07-11
Began adding a commit for a lock contention logging benchmark but deferring to doing that separately.