@fanquake 2022-07-20
~0 - I'm not really convinced we need more benchmarks for logging code.