@jonatack 2022-07-20
The logging is migrating to use these methods; it makes sense to benchmark them, or to remove the existing logging benchmarks.