@jonatack 2022-08-05
This change makes sense given that the codebase is migrating to these methods but closing for now, will try again later.